modified:   api/server/handler/errors.go modified:   api/server/handler/order.go modified:   storage/psql/migrations/initial.up.sql modified:   storage/psql/service.go

+++ storage/psql/migrations/initial.up.sql
@@ -11,7 +11,7 @@ CREATE TABLE IF NOT EXISTS users
 
 CREATE TABLE IF NOT EXISTS orders
 (
-    num int not null,
+    num bigint not null,
 	status int not null,
 	accrual int,
 	uploaded_at timestamp    not null default now(),
